Fixed test autoloading and split file containing two classes.
