+++ phpunit.xml.dist
@@ -1,7 +1,7 @@
 <phpunit bootstrap="./tests/Bootstrap.php" colors="true">
     <testsuites>
         <testsuite name="ZendService Amazon Test Suite">
-            <directory>./tests/ZendService</directory>
+            <directory>./tests/ZendServiceTest</directory>
         </testsuite>
     </testsuites>
