+++ tests/ZendServiceTest/Amazon/Authentication/S3Test.php
