+++ tests/ZendServiceTest/Amazon/Authentication/V1Test.php
