+++ tests/ZendServiceTest/Amazon/Authentication/V2Test.php
