+++ tests/ZendServiceTest/Amazon/Authentication/_files/ec2_v2_return.txt
