+++ tests/ZendServiceTest/Amazon/Authentication/_files/sqs_v2_get_return.txt
