+++ tests/ZendServiceTest/Amazon/Ec2/AvailabilityZonesTest.php
