+++ tests/ZendServiceTest/Amazon/Ec2/CloudWatchTest.php
