+++ tests/ZendServiceTest/Amazon/Ec2/EbsTest.php
