+++ tests/ZendServiceTest/Amazon/Ec2/Ec2Test.php
