+++ tests/ZendServiceTest/Amazon/Ec2/ElasticIpTest.php
