+++ tests/ZendServiceTest/Amazon/Ec2/ImageTest.php
