+++ tests/ZendServiceTest/Amazon/Ec2/InstanceReservedTest.php
