+++ tests/ZendServiceTest/Amazon/Ec2/InstanceTest.php
