+++ tests/ZendServiceTest/Amazon/Ec2/InstanceWindowsTest.php
