+++ tests/ZendServiceTest/Amazon/Ec2/KeypairTest.php
