+++ tests/ZendServiceTest/Amazon/Ec2/RegionTest.php
