+++ tests/ZendServiceTest/Amazon/Ec2/SecurityGroupsTest.php
