+++ tests/ZendServiceTest/Amazon/Ec2/TestAmazonAbstract.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * Zend Framework (http://framework.zend.com/)
+ *
+ * @link      http://github.com/zendframework/zf2 for the canonical source repository
+ * @copyright Copyright (c) 2005-2014 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @package   Zend_Service
+ */
+
+namespace ZendServiceTest\Amazon\Ec2;
+
+class TestAmazonAbstract extends \ZendService\Amazon\Ec2\AbstractEc2
+{
+
+    public function returnRegion()
+    {
+        return $this->_region;
+    }
+
+    public function testSign($params)
+    {
+        return $this->signParameters($params);
+    }
+}
