+++ tests/ZendServiceTest/Amazon/OfflineTest.php
