+++ tests/ZendServiceTest/Amazon/S3/OfflineStreamTest.php
