+++ tests/ZendServiceTest/Amazon/S3/OfflineTest.php
