+++ tests/ZendServiceTest/Amazon/S3/OnlineTest.php
