+++ tests/ZendServiceTest/Amazon/S3/S3RestTest.php
