+++ tests/ZendServiceTest/Amazon/S3/StreamTest.php
