+++ tests/ZendServiceTest/Amazon/Sqs/OnlineTest.php
