+++ tests/ZendServiceTest/Amazon/_files/amazon-response-invalid.xml
