+++ tests/ZendServiceTest/Amazon/_files/amazon-response-request-throttled-error.xml
