+++ tests/ZendServiceTest/Amazon/_files/amazon-response-valid.xml
