+++ tests/ZendServiceTest/Amazon/_files/offers_with_names.xml
